backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-dateutil==2.9.0
python-jose[cryptography]==3.3.0
httpx>=0.27.0
